Move Puzzle struct/methods to internal sudoku pkg
